// src/app/services/model/budgetModel.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BudgetModel { 
    id?: number | null;
    amount?: number;
    month?: string;
    categoryId?: number | null;
}